Desenvolvendo elementos com distinct
