+++ kotlin-collections-projeto-inicial/src/br/com/alura/main.kt
@@ -1,39 +1,10 @@
 package br.com.alura
 
 fun main() {
-    val banco = BancoDeNomes()
-    val nomesSalvos: Collection<String> = banco.nomes
-    banco.salva("Alex")
-    println(nomesSalvos)
-    println(BancoDeNomes().nomes)
-}
-
-class BancoDeNomes {
-
-    val nomes: Collection<String> get() = dados.toList()
-
-    fun salva(nome: String) {
-        dados.add(nome)
-    }
-
-    companion object {
-        private val dados = mutableListOf<String>()
-    }
-}
-
-fun testaColecao() {
-    val nomes: Collection<String> = mutableListOf(
-        "Alex",
-        "Fran",
-        "Gui",
-        "Maria",
-        "Ana"
-    )
-
-    for (nome in nomes.iterator()) {
-        println(nome)
-    }
-    println(nomes)
-    println("Tem o nome Alex? ${nomes.contains("Vinicius")}")
-    println("Tamanho da coleção: ${nomes.size}")
+    val assistiramCursoAndroid = listOf("Alex", "Fran", "Gui", "Maria")
+    val assistiramCursoKotlin = listOf("Alex", "Paulo", "Maria")
+    val assistiramAmbos = mutableListOf<String>()
+    assistiramAmbos.addAll(assistiramCursoAndroid)
+    assistiramAmbos.addAll(assistiramCursoKotlin)
+    println(assistiramAmbos.distinct())
 }
